requirements.txt
streamlit==1.38.0
pandas
numpy
scikit-learn
matplotlib
seaborn
opencv-python-headless==4.10.0.84
opencv-contrib-python-headless==4.10.0.84
streamlit-webrtc==0.47.7
av==10.0.0
pillow
